Replace SplitterLayout with a new ResizableWithSplitter component.
This replacement has the following advantages:
It's compatible with React 18, and it supports making "max-height"
the controlled property, which is useful for our timeline splitter.

In the past, we needed to compute the height of the timeline contents,
and then make it so you couldn't move the splitter down more than the
full timeline height, because we didn't want there to be a gap above
the splitter, or even initialize the splitter at a position that creates
such a gap. By making the splitter control max-height instead, even if
the splitter is dragged low enough to say max-height:1000px, the splitter
won't actually move that low because the timeline won't grow beyond its
full size (because there's no longer a height property forcing it to grow).