Replace symbolicator-cli with a profiler-edit node tool
Before:

```
  node dist/symbolicator-cli.js \
    --input samply-profile.json \
    --output profile-symbolicated.json \
    --server http://localhost:3000
```

After:

```
  node node-tools-dist/profiler-edit.js \
    -i samply-profile.json \
    -o profile-symbolicated.json \
    --symbolicate-with-server http://localhost:3000
```

The new tool also accepts a URL or profile storage hash as input, and can
write gzip-compressed output:

```
  node node-tools-dist/profiler-edit.js \
    --from-hash w1spyw917hg... \
    -o out.json.gz \
    --symbolicate-with-server http://localhost:8001/abcdef/
```

The reason for this change is that I want to add further ways to
modify profiles in the future (such as an `--insert-label-frames`
argument) and it doesn't really make sense to have a separate tool
for each modification.

This commit also moves the script file to /src/node-tools/ and the output to /node-tools-dist/.
I'm planning to add more scripts of this type in the future.